+++ App.Core.ApplicationService/ApplicationSerrvices/Countries/CountryService.cs
@@ -1,5 +1,6 @@
 ﻿using App.Core.ApplicationService.Dtos.CountryDtos;
 using App.Core.ApplicationService.IRepositories;
+using App.Core.Entities.Exceptions;
 using App.Core.Entities.Model;
 using AutoMapper;
 using System;
@@ -39,12 +40,21 @@ namespace App.Core.ApplicationService.ApplicationSerrvices.Countries
 
         public int Delete(int id)
         {
+            if (countryRepository.GetQuery().Select(x => x.Id != id).FirstOrDefault())
+            {
+                throw new InvalidIdException("Wrong Id");
+            }
             countryRepository.Delete(id);
             return id;
         }
 
         public async Task<Country> Get(int id)
         {
+            if (countryRepository.GetQuery().Select(x => x.Id != id).FirstOrDefault())
+            {
+                throw new InvalidIdException("Wrong Id");
+            }
+
             return await countryRepository.Get(id);
         }
